+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog of threedigrid-builder
 
 - Added 2D boundary conditions.
 
+- Enable compression in HDF5 output.
+
 - Fixed 2D lines that connect a larger to a smaller cell in south east direction.